Make pretty printer generic over languages. Make all languages consistently use DeBruijnIndex instead of usize iside Var.
