+++ src/bin/smiley/lang.rs
@@ -134,7 +134,7 @@ impl Teachable for Smiley {
         match binding_expr {
             BindingExpr::Lambda(body) => AstNode::new(Self::Lambda, [body]),
             BindingExpr::Apply(fun, arg) => AstNode::new(Self::Apply, [fun, arg]),
-            BindingExpr::Var(index) => AstNode::leaf(Self::Var(DeBruijnIndex(index))),
+            BindingExpr::Var(index) => AstNode::leaf(Self::Var(index)),
             BindingExpr::LibVar(ix) => AstNode::leaf(Self::LibVar(ix)),
             BindingExpr::Lib(ix, bound_value, body) => {
                 AstNode::new(Self::Lib(ix), [bound_value, body])
@@ -147,7 +147,7 @@ impl Teachable for Smiley {
         let binding_expr = match node.as_parts() {
             (Self::Lambda, [body]) => BindingExpr::Lambda(body),
             (Self::Apply, [fun, arg]) => BindingExpr::Apply(fun, arg),
-            (&Self::Var(DeBruijnIndex(index)), []) => BindingExpr::Var(index),
+            (&Self::Var(index), []) => BindingExpr::Var(index),
             (Self::Lib(ix), [bound_value, body]) => BindingExpr::Lib(*ix, bound_value, body),
             (Self::Shift, [body]) => BindingExpr::Shift(body),
             _ => return None,
